+++ src/r_voxel.c
@@ -1075,8 +1075,8 @@ static void DrawColumnCubesLoop8(
 	const fixed_t ux2,
 	const boolean do_voxel_radial_fog
 ) {
-	const int linesize = video.width;
-	pixel_t * dest = I_VideoBuffer + viewwindowy * linesize + viewwindowx;
+	const int linesize = video.height;
+	pixel_t * dest = I_VideoBuffer + (viewwindowx * linesize) + viewwindowy;
 
 	const lighttable_t *thiscolormap =
 		(spr->tint >= 0) ? colormaps[spr->tint] : fullcolormap;
@@ -1183,7 +1183,7 @@ static void DrawColumnCubesLoop8(
 
 				for (; uy < uy1 ; uy += FRACUNIT)
 				{
-					dest[(uy >> FRACBITS) * linesize + (ux >> FRACBITS)] = pix;
+					dest[(ux >> FRACBITS) * linesize + (uy >> FRACBITS)] = pix;
 				}
 			}
 			else if (has_bottom)
@@ -1198,7 +1198,7 @@ static void DrawColumnCubesLoop8(
 
 				for (; uy > uy2 ; uy -= FRACUNIT)
 				{
-					dest[(uy >> FRACBITS) * linesize + (ux >> FRACBITS)] = pix;
+					dest[(ux >> FRACBITS) * linesize + (uy >> FRACBITS)] = pix;
 				}
 			}
 
@@ -1216,7 +1216,7 @@ static void DrawColumnCubesLoop8(
 					byte src = slab[i];
 					pixel_t pix = colormap[spr->brightmap[src]][dc_translation[src]]; // [Nugget] Translation
 
-					dest[(uy >> FRACBITS) * linesize + (ux >> FRACBITS)] = pix;
+					dest[(ux >> FRACBITS) * linesize + (uy >> FRACBITS)] = pix;
 				}
 			}
 		}
@@ -1243,8 +1243,8 @@ static void DrawColumnCubesLoop32(
 	const fixed_t ux2,
 	const boolean do_voxel_radial_fog
 ) {
-	const int linesize = video.width;
-	pixel32_t * dest = I_VideoBuffer32 + viewwindowy * linesize + viewwindowx;
+	const int linesize = video.height;
+	pixel32_t * dest = I_VideoBuffer32 + (viewwindowx * linesize) + viewwindowy;
 
 	const lighttable32_t *thiscolormap =
 		(spr->tint >= 0) ? colormaps32[spr->tint] : fullcolormap32;
@@ -1348,7 +1348,7 @@ static void DrawColumnCubesLoop32(
 
 				for (; uy < uy1 ; uy += FRACUNIT)
 				{
-					dest[(uy >> FRACBITS) * linesize + (ux >> FRACBITS)] = pix;
+					dest[(ux >> FRACBITS) * linesize + (uy >> FRACBITS)] = pix;
 				}
 			}
 			else if (has_bottom)
@@ -1363,7 +1363,7 @@ static void DrawColumnCubesLoop32(
 
 				for (; uy > uy2 ; uy -= FRACUNIT)
 				{
-					dest[(uy >> FRACBITS) * linesize + (ux >> FRACBITS)] = pix;
+					dest[(ux >> FRACBITS) * linesize + (uy >> FRACBITS)] = pix;
 				}
 			}
 
@@ -1381,7 +1381,7 @@ static void DrawColumnCubesLoop32(
 					byte src = slab[i];
 					pixel32_t pix = colormap[spr->brightmap[src]][dc_translation[src]]; // [Nugget] Translation
 
-					dest[(uy >> FRACBITS) * linesize + (ux >> FRACBITS)] = pix;
+					dest[(ux >> FRACBITS) * linesize + (uy >> FRACBITS)] = pix;
 				}
 			}
 		}
@@ -1410,8 +1410,8 @@ static void DrawColumnCubesLoopDithered8(
 	const fixed_t ux2,
 	const boolean do_voxel_radial_fog
 ) {
-	const int linesize = video.width;
-	pixel_t * dest = I_VideoBuffer + viewwindowy * linesize + viewwindowx;
+	const int linesize = video.height;
+	pixel_t * dest = I_VideoBuffer + (viewwindowx * linesize) + viewwindowy;
 
 	const lighttable_t *thiscolormap =
 		(spr->tint >= 0) ? colormaps[spr->tint] : fullcolormap;
@@ -1534,7 +1534,7 @@ static void DrawColumnCubesLoopDithered8(
 
 				for (; uy < uy1 ; uy += FRACUNIT)
 				{
-					dest[(uy >> FRACBITS) * linesize + (ux >> FRACBITS)] = pix[dither_pattern_row[dx]];
+					dest[(ux >> FRACBITS) * linesize + (uy >> FRACBITS)] = pix[dither_pattern_row[dx]];
 					dx = (dx + 1) & DITHER_PATTERN_WIDTH_MASK;
 				}
 			}
@@ -1555,7 +1555,7 @@ static void DrawColumnCubesLoopDithered8(
 
 				for (; uy > uy2 ; uy -= FRACUNIT)
 				{
-					dest[(uy >> FRACBITS) * linesize + (ux >> FRACBITS)] = pix[dither_pattern_row[dx]];
+					dest[(ux >> FRACBITS) * linesize + (uy >> FRACBITS)] = pix[dither_pattern_row[dx]];
 					dx = (dx + 1) & DITHER_PATTERN_WIDTH_MASK;
 				}
 			}
@@ -1575,7 +1575,7 @@ static void DrawColumnCubesLoopDithered8(
 					const byte src = slab[i];
 					const pixel_t pix = colormap[dither_pattern_row[dx]][spr->brightmap[src]][dc_translation[src]];
 
-					dest[(uy >> FRACBITS) * linesize + (ux >> FRACBITS)] = pix;
+					dest[(ux >> FRACBITS) * linesize + (uy >> FRACBITS)] = pix;
 					dx = (dx + 1) & DITHER_PATTERN_WIDTH_MASK;
 				}
 			}
@@ -1603,8 +1603,8 @@ static void DrawColumnCubesLoopDithered32(
 	const fixed_t ux2,
 	const boolean do_voxel_radial_fog
 ) {
-	const int linesize = video.width;
-	pixel32_t * dest = I_VideoBuffer32 + viewwindowy * linesize + viewwindowx;
+	const int linesize = video.height;
+	pixel32_t * dest = I_VideoBuffer32 + (viewwindowx * linesize) + viewwindowy;
 
 	const lighttable32_t *thiscolormap =
 		(spr->tint >= 0) ? colormaps32[spr->tint] : fullcolormap32;
@@ -1727,7 +1727,7 @@ static void DrawColumnCubesLoopDithered32(
 
 				for (; uy < uy1 ; uy += FRACUNIT)
 				{
-					dest[(uy >> FRACBITS) * linesize + (ux >> FRACBITS)] = pix[dither_pattern_row[dx]];
+					dest[(ux >> FRACBITS) * linesize + (uy >> FRACBITS)] = pix[dither_pattern_row[dx]];
 					dx = (dx + 1) & DITHER_PATTERN_WIDTH_MASK;
 				}
 			}
@@ -1748,7 +1748,7 @@ static void DrawColumnCubesLoopDithered32(
 
 				for (; uy > uy2 ; uy -= FRACUNIT)
 				{
-					dest[(uy >> FRACBITS) * linesize + (ux >> FRACBITS)] = pix[dither_pattern_row[dx]];
+					dest[(ux >> FRACBITS) * linesize + (uy >> FRACBITS)] = pix[dither_pattern_row[dx]];
 					dx = (dx + 1) & DITHER_PATTERN_WIDTH_MASK;
 				}
 			}
@@ -1768,7 +1768,7 @@ static void DrawColumnCubesLoopDithered32(
 					const byte src = slab[i];
 					const pixel32_t pix = colormap[dither_pattern_row[dx]][spr->brightmap[src]][dc_translation[src]];
 
-					dest[(uy >> FRACBITS) * linesize + (ux >> FRACBITS)] = pix;
+					dest[(ux >> FRACBITS) * linesize + (uy >> FRACBITS)] = pix;
 					dx = (dx + 1) & DITHER_PATTERN_WIDTH_MASK;
 				}
 			}
@@ -1986,8 +1986,8 @@ static void DrawColumnBoundedLoop8(
 	const fixed_t ux2,
 	const boolean do_voxel_radial_fog
 ) {
-	const int linesize = video.width;
-	pixel_t *dest = I_VideoBuffer + viewwindowy * linesize + viewwindowx;
+	const int linesize = video.height;
+	pixel_t *dest = I_VideoBuffer + (viewwindowx * linesize) + viewwindowy;
 
 	const lighttable_t *thiscolormap =
 		(spr->tint >= 0) ? colormaps[spr->tint] : fullcolormap;
@@ -2001,9 +2001,9 @@ static void DrawColumnBoundedLoop8(
 	              x2 = (spr->x2 + 1) << FRACBITS;
 
 	int uxi = ux >> FRACBITS;
-	dest += uxi;
+	dest += uxi * linesize;
 
-	for (; ux < ux2;  ux += FRACUNIT, uxi++, dest++)
+	for (; ux < ux2;  ux += FRACUNIT, uxi++, dest += linesize)
 	{
 		if (ux >= x2) break;
 		if (ux <  x1) continue;
@@ -2061,9 +2061,9 @@ static void DrawColumnBoundedLoop8(
 			}
 
 			fixed_t uy = ((uy1 - 1) | FRACMASK) + 1;
-			pixel_t *dest2 = dest + (uy >> FRACBITS) * linesize;
+			pixel_t *dest2 = dest + (uy >> FRACBITS);
 
-			for (; uy <= uy2;  uy += FRACUNIT, dest2 += linesize)
+			for (; uy <= uy2;  uy += FRACUNIT, dest2++)
 			{
 				int i = (((uy - uy0) >> FRACBITS) * imidscale) >> FRACBITS;
 
@@ -2093,8 +2093,8 @@ static void DrawColumnBoundedLoop32(
 	const fixed_t ux2,
 	const boolean do_voxel_radial_fog
 ) {
-	const int linesize = video.width;
-	pixel32_t *dest = I_VideoBuffer32 + viewwindowy * linesize + viewwindowx;
+	const int linesize = video.height;
+	pixel32_t *dest = I_VideoBuffer32 + (viewwindowx * linesize) + viewwindowy;
 
 	const lighttable32_t *thiscolormap =
 		(spr->tint >= 0) ? colormaps32[spr->tint] : fullcolormap32;
@@ -2108,9 +2108,9 @@ static void DrawColumnBoundedLoop32(
 	              x2 = (spr->x2 + 1) << FRACBITS;
 
 	int uxi = ux >> FRACBITS;
-	dest += uxi;
+	dest += uxi * linesize;
 
-	for (; ux < ux2;  ux += FRACUNIT, uxi++, dest++)
+	for (; ux < ux2;  ux += FRACUNIT, uxi++, dest += linesize)
 	{
 		if (ux >= x2) break;
 		if (ux <  x1) continue;
@@ -2168,9 +2168,9 @@ static void DrawColumnBoundedLoop32(
 			}
 
 			fixed_t uy = ((uy1 - 1) | FRACMASK) + 1;
-			pixel32_t *dest2 = dest + (uy >> FRACBITS) * linesize;
+			pixel32_t *dest2 = dest + (uy >> FRACBITS);
 
-			for (; uy <= uy2;  uy += FRACUNIT, dest2 += linesize)
+			for (; uy <= uy2;  uy += FRACUNIT, dest2++)
 			{
 				int i = (((uy - uy0) >> FRACBITS) * imidscale) >> FRACBITS;
 
@@ -2202,8 +2202,8 @@ static void DrawColumnBoundedLoopDithered8(
 	const fixed_t ux2,
 	const boolean do_voxel_radial_fog
 ) {
-	const int linesize = video.width;
-	pixel_t *dest = I_VideoBuffer + viewwindowy * linesize + viewwindowx;
+	const int linesize = video.height;
+	pixel_t *dest = I_VideoBuffer + (viewwindowx * linesize) + viewwindowy;
 
 	const lighttable_t *thiscolormap =
 		(spr->tint >= 0) ? colormaps[spr->tint] : fullcolormap;
@@ -2219,9 +2219,9 @@ static void DrawColumnBoundedLoopDithered8(
 	              x2 = (spr->x2 + 1) << FRACBITS;
 
 	int uxi = ux >> FRACBITS;
-	dest += uxi;
+	dest += uxi * linesize;
 
-	for (; ux < ux2;  ux += FRACUNIT, uxi++, dest++)
+	for (; ux < ux2;  ux += FRACUNIT, uxi++, dest += linesize)
 	{
 		if (ux >= x2) break;
 		if (ux <  x1) continue;
@@ -2291,11 +2291,11 @@ static void DrawColumnBoundedLoopDithered8(
 			}
 
 			fixed_t uy = ((uy1 - 1) | FRACMASK) + 1;
-			pixel_t *dest2 = dest + (uy >> FRACBITS) * linesize;
+			pixel_t *dest2 = dest + (uy >> FRACBITS);
 
 			byte dx = (uy >> FRACBITS) & DITHER_PATTERN_WIDTH_MASK;
 
-			for (; uy <= uy2;  uy += FRACUNIT, dest2 += linesize)
+			for (; uy <= uy2;  uy += FRACUNIT, dest2++)
 			{
 				int i = (((uy - uy0) >> FRACBITS) * imidscale) >> FRACBITS;
 
@@ -2326,8 +2326,8 @@ static void DrawColumnBoundedLoopDithered32(
 	const fixed_t ux2,
 	const boolean do_voxel_radial_fog
 ) {
-	const int linesize = video.width;
-	pixel32_t *dest = I_VideoBuffer32 + viewwindowy * linesize + viewwindowx;
+	const int linesize = video.height;
+	pixel32_t *dest = I_VideoBuffer32 + (viewwindowx * linesize) + viewwindowy;
 
 	const lighttable32_t *thiscolormap =
 		(spr->tint >= 0) ? colormaps32[spr->tint] : fullcolormap32;
@@ -2343,9 +2343,9 @@ static void DrawColumnBoundedLoopDithered32(
 	              x2 = (spr->x2 + 1) << FRACBITS;
 
 	int uxi = ux >> FRACBITS;
-	dest += uxi;
+	dest += uxi * linesize;
 
-	for (; ux < ux2;  ux += FRACUNIT, uxi++, dest++)
+	for (; ux < ux2;  ux += FRACUNIT, uxi++, dest += linesize)
 	{
 		if (ux >= x2) break;
 		if (ux <  x1) continue;
@@ -2415,11 +2415,11 @@ static void DrawColumnBoundedLoopDithered32(
 			}
 
 			fixed_t uy = ((uy1 - 1) | FRACMASK) + 1;
-			pixel32_t *dest2 = dest + (uy >> FRACBITS) * linesize;
+			pixel32_t *dest2 = dest + (uy >> FRACBITS);
 
 			byte dx = (uy >> FRACBITS) & DITHER_PATTERN_WIDTH_MASK;
 
-			for (; uy <= uy2;  uy += FRACUNIT, dest2 += linesize)
+			for (; uy <= uy2;  uy += FRACUNIT, dest2++)
 			{
 				int i = (((uy - uy0) >> FRACBITS) * imidscale) >> FRACBITS;