+++ src/v_video.h
@@ -173,7 +173,7 @@ void V_UseBuffer32(pixel32_t *buffer, int pitch);
 void V_RestoreBuffer(void);
 
 void V_CopyRect(int srcx, int srcy, pixel_t *source, int width, int height,
-                int destx, int desty);
+                int pitch, int destx, int desty);
 
 void V_CopyRect32(int srcx, int srcy, pixel32_t *source, int width, int height,
                   int destx, int desty);