+++ CHANGELOG.md
@@ -5,6 +5,7 @@ None.
 ## Changes
 
 - **Merged changes from Woof! post-16.0.0**
+- **Improved low-resolution post-process**
 
 ## Bug Fixes